requirements.txt
keras
plaidml-keras
matplotlib
Pillow
pandas
receptivefield==0.4.0
seaborn
keras-vis
scipy==1.2.1